refactor: :recycle: A little more of refactoring
