+++ src/components/business/index.ts
@@ -1,5 +1,5 @@
 import BusinessContainer from './BusinessContainer'
 import BusinessCard from './BusinessCard'
-import BusinessMapMarker from './BusinessMapMarker'
+import BusinessMapMarker from './MapMarker'
 
 export { BusinessContainer, BusinessCard, BusinessMapMarker }
